Update standard for networkv2 before overcloud
Apply ideas described in the following to the standard
overcloud deployment:

  https://github.com/fultonj/xena/tree/main/networkv2

Assumption is that after tripleo-lab is run first and
that lab_pickup.sh is run after that to create the
~/vip_subnet_map.yaml and then copy in the deployed
metal and network files which were already created
but to change their path to use ~/templates prior.

Subsequent deployments will use the metalsmith directory
and it will create the metal and network files based on
the same convention.

Either way the standard/deploy.sh is following the
networkv2 pattern seen in networkv2/deploy.sh.