In metalsmith/provision.sh default network to false
Because tripleo-lab will provision the network and it
doesn't need to be reprovioned per deployment (though
the metal does) do not provision network for every
metalsmith deployment by default.

Either way ensure the $NETWORK_OUTPUT_FILE is in the
metalsmith directory so that the deploy script will
copy it into the deployment directory (e.g. hci or
standard) when they copy in $METAL_OUTPUT_FILE.