deployed_ceph script use same ceph container tag as undercloud
The stanard deployment effecitvely uses this:

  --container-namespace undercloud.ctlplane.mydomain.tld:8787/ceph
  --container-image daemon \
  --container-tag v6.0.4-stable-6.0-pacific-centos-8-x86_64 \

becuase my overcloud is configured at that point to accept as a regsitry
undercloud.ctlplane.mydomain.tld. However, at this point of the deployment
the overcloud is not configured to accept the undercloud as a registry. So
I will use the following which is the soruce of where my undercloud got it:

  --container-namespace quay.io/ceph \