Temporarily don't set hostname
cephadm would fail to apply the spec with this error:

Error EINVAL: Host standalone (192.168.24.2) failed check(s): ['hostname "standalone.localdomain" does not match expected hostname "standalone"']

regardless of if the spec had standalone.localdomain or
standalone.

---
    if 'expect_hostname' in ctx and ctx.expect_hostname:
        if get_hostname().lower() != ctx.expect_hostname.lower():
            errors.append('hostname "%s" does not match expected hostname "%s"' % (
                get_hostname(), ctx.expect_hostname))
        logger.info('Hostname "%s" matches what is expected.',
                    ctx.expect_hostname)
---

Problem is:
- hostname is getting set to "standalone.localdomain"
- ctx.hostname is getting set to "standalone"

Not setting the hostname this way is a workaround to get ceph stood up.
We can then set the hostname as standalone requires before we run
standalone.

This is a mess I'll have to clean up later. For now I continue with POC.