Standalone deployed_ceph on 'storage' network
TripleO standalone needs to configure the 192.168.24.0/24 network
and interface so it doesn't exist for me to use it with deployed
ceph.

Because my VM already has an IP I modified the deployed_ceph deploy
to use that IP instead and then I'll try passing it to the standalone
deployment as if it were a separate storage network.