Use initial_ceph.conf in standalone playbook
As per standlone doc [1] use an initial_ceph.conf.
Though Ceph will be configured to run on a single host
via the --single-host-defaults option, this deployment
only has a single OSD so it cannot replicate data even
on the same host.

Passing the initial conf also has an important side
effect of not setting public/cluster networks which
standlone needs as ctlplane network is default but
it is not yet configured. This results in mgr not
starting and failing with -1 unable to find any IP
address in networks '192.168.24.0/24'. That happens
since the assimilate ceph conf will have the same
public network = 192.168.24.0/24 = cluster network
because of [2] [3] [4].

[1] https://docs.openstack.org/project-deploy-guide/tripleo-docs/latest/deployment/standalone.html

[2] https://github.com/openstack/tripleo-ansible/blob/master/tripleo_ansible/roles/tripleo_cephadm/templates/ceph.conf.j2#L15-L20

[3] https://github.com/openstack/python-tripleoclient/blob/master/tripleoclient/v2/overcloud_ceph.py#L372-L376

[4] https://github.com/openstack/python-tripleoclient/blob/master/tripleoclient/utils.py#L3282-L3283